splitting AutoSSO helpers into another file for easier unit testing
