+++ src/platform/site-wide/user-nav/containers/AutoSSO.jsx
@@ -1,6 +1,4 @@
-import React from 'react';
 import { connect } from 'react-redux';
-import { fromPairs } from 'lodash';
 
 import { isLoggedIn } from 'platform/user/selectors';
 import { checkKeepAlive } from 'platform/user/authentication/actions';
@@ -9,59 +7,26 @@ import {
   ssoeInbound,
   hasCheckedKeepAlive,
 } from 'platform/user/authentication/selectors';
-import { autoLogin, autoLogout } from 'platform/user/authentication/utilities';
-import { hasSession, hasSessionSSO } from 'platform/user/profile/utilities';
-import {
-  ssoKeepAliveSession,
-  getForceAuth,
-  setForceAuth,
-  deleteForceAuth,
-} from 'platform/utilities/sso';
+import { setForceAuth, removeForceAuth } from 'platform/utilities/sso';
 
-function parseqs(value) {
-  /*
-   * naive query string parsing function, takes a query string and returns
-   * an object mapping the keys to values.
-   */
-  const data = value.startsWith('?') ? value.substring(1) : value;
-  const entries = data ? data.split('&').map(q => q.split('=')) : [];
-  return fromPairs(entries);
-}
+import { parseqs, checkAutoSession } from '../helpers';
 
-export async function checkStatus(toggleKeepAlive) {
-  await ssoKeepAliveSession();
-  if (hasSession() && !hasSessionSSO()) {
-    autoLogout();
-  } else if (!hasSession() && hasSessionSSO() && !getForceAuth()) {
-    autoLogin();
+function AutoSSO(props) {
+  const { useSSOe, useInboundSSOe, hasCalledKeepAlive, userLoggedIn } = props;
+
+  if (parseqs(window.location.search).auth === 'force-needed') {
+    setForceAuth();
   }
 
-  toggleKeepAlive();
-}
-
-class AutoSSO extends React.Component {
-  render() {
-    const {
-      useSSOe,
-      useInboundSSOe,
-      hasCalledKeepAlive,
-      userLoggedIn,
-    } = this.props;
-
-    if (parseqs(window.location.search).auth === 'force-needed') {
-      setForceAuth();
-    }
-
-    if (userLoggedIn) {
-      deleteForceAuth();
-    }
-
-    if (useSSOe && useInboundSSOe && !hasCalledKeepAlive) {
-      checkStatus(this.props.checkKeepAlive);
-    }
-
-    return null;
+  if (userLoggedIn) {
+    removeForceAuth();
   }
+
+  if (useSSOe && useInboundSSOe && !hasCalledKeepAlive) {
+    checkAutoSession(props.checkKeepAlive);
+  }
+
+  return null;
 }
 
 const mapStateToProps = state => ({
